refactor(ui): extract request response view
